Keep reading from same fd after binary check
Previously the fd was closed and the entire file was re-read
after the binary check was done.

Now the fd is kept open and the remaining file is read into a new
buffer that is then joined with the binary check buffer and converted
to a string.

This prevents re-reading the first 512 bytes per file read.